enregistrements de la reponse dans un fichier de sortie avec les elements requis nom du graph shortestPath expectedpath hallucination et  score

+++ appli/src/main/java/com/example/springiapromptdemo/entities/GraphDatasetElement.java
@@ -14,7 +14,7 @@ public class GraphDatasetElement {
     private Long id;
     private String pointA;
     private String pointB;
-    private Float distance;
+    private Double distance;
 
     private String grapheName;
